+++ python/cugraph/tests/test_sssp.py
@@ -17,10 +17,9 @@ import time
 
 import numpy as np
 import pytest
-from scipy.io import mmread
 
-import cudf
 import cugraph
+from cugraph.tests import utils
 from librmm_cffi import librmm as rmm
 from librmm_cffi import librmm_config as rmm_cfg
 
@@ -36,17 +35,6 @@ with warnings.catch_warnings():
 
 
 print('Networkx version : {} '.format(nx.__version__))
-
-
-def read_mtx_file(mm_file):
-    print('Reading ' + str(mm_file) + '...')
-    return mmread(mm_file).asfptype()
-
-
-def read_csv_file(mm_file):
-    print('Reading ' + str(mm_file) + '...')
-    return cudf.read_csv(mm_file, delimiter=' ',
-                         dtype=['int32', 'int32', 'float32'], header=None)
 
 
 def cugraph_call(cu_M, source, edgevals=False):
@@ -131,8 +119,8 @@ def test_sssp(managed, pool, graph_file, source):
 
     assert(rmm.is_initialized())
 
-    M = read_mtx_file(graph_file+'.mtx')
-    cu_M = read_csv_file(graph_file+'.csv')
+    M = utils.read_mtx_file(graph_file+'.mtx')
+    cu_M = utils.read_csv_file(graph_file+'.csv')
     cu_paths = cugraph_call(cu_M, source)
     nx_paths, Gnx = networkx_call(M, source)
 
@@ -168,8 +156,8 @@ def test_sssp_edgevals(managed, pool, graph_file, source):
 
     assert(rmm.is_initialized())
 
-    M = read_mtx_file(graph_file+'.mtx')
-    cu_M = read_csv_file(graph_file+'.csv')
+    M = utils.read_mtx_file(graph_file+'.mtx')
+    cu_M = utils.read_csv_file(graph_file+'.csv')
     cu_paths = cugraph_call(cu_M, source, edgevals=True)
     nx_paths, Gnx = networkx_call(M, source, edgevals=True)
